minor adjustments to constructor
docs specified double and not int as arg type.

+++ creditCard.cs
@@ -8,7 +8,7 @@ namespace Assignment_5
 {
     class CreditCard: Payment
     {
-        public CreditCard(int amount, string ccnum, string cctype, string expdate, string ccholdername, string purchaseDate)
+        public CreditCard(double amount, string ccnum, string cctype, string expdate, string ccholdername, string purchaseDate)
         {
             CCNum = ccnum;
             expDate = expdate;
